Create a grid

create.grid allows to create a new grid or modify an existing grid that is used by fit.momo to predict movement rates and required for the matrix exponential approach.

Usage

create.grid(
  xrange = c(0, 1),
  yrange = c(0, 1),
  dxdy = c(0.1, 0.1),
  select = FALSE,
  plot.land = FALSE,
  keep.gpar = FALSE,
  verbose = TRUE,
  fit = NULL,
  grid = NULL
)

Arguments

xrange

range of the x-dimension of spatial domain. Default: c(0,1).

yrange

range of the y-dimension of spatial domain. Default: c(0,1).

dxdy

resolution of grid in x and y direction. Default: c(0.1,0.1).

select

logical; if TRUE, allows to select cells in spatial grid. Default: FALSE.

plot.land

plot.land logical; If TRUE, plot land masses using the function maps::map. Default: FALSE.

keep.gpar

logical; If TRUE, do not overwrite the graphical parameters. Default: FALSE.

verbose

if TRUE, print information to console. Default: TRUE.

fit

optional; allows to extract and update grid from fitted object of class momo.fit returned by function fit.momo. Default: NULL. NULL.

grid

optional; allows to update a grid of class momo.grid as returned by the function create.grid. Default: NULL.

Value

A list of class momo.grid with a information about the spatial domain and grid.

Examples

grid <- create.grid()