Fit the movement model

The movement model, momo, allows estimation of habitat preferences and fine-scale movement rates using various types of tagging data. It supports flexible spatiotemporal modeling with options for Kalman filter or matrix exponential approaches, and includes tools for data preparation, model fitting, and visualization.

Usage

fit.momo(
  dat,
  conf = NULL,
  par = NULL,
  map = NULL,
  run = TRUE,
  lower = get.lower.bounds(par),
  upper = get.upper.bounds(par),
  rel.tol = 1e-10,
  do.sdreport = TRUE,
  do.report = TRUE,
  use.expm = NULL,
  use.rel.events = NULL,
  dbg = FALSE,
  control = NULL,
  verbose = TRUE,
  ...
)

Arguments

dat

data frame with input data as produced by the function check.momo.data.

conf

configuration list as produced by the function def.conf. If NULL (default), def.conf is used to generate the default configuration list based on the input data.

par

parameter list with initial values as produced by the function def.par. If NULL (default), def.par is used to generate the default parameter list based on the input data and configuration list.

map

list with parameter mapping as produced by the function def.map. If NULL (default), def.map is used to generate the default mapping list based on the input data and configuration and parameter lists.

run

logical; If FALSE, the AD object is returned with the optimization. Default: FALSE.

lower

by default, get.lower.bounds is used.

upper

by default, get.upper.bounds is used.

rel.tol

option passed to stats::nlminb sets the convergence criteria. Default: 1e-10.

do.sdreport

logical; If FALSE, RTMB::sdreport is not run and no parameter uncertainties are returned. Default: TRUE.

do.report

logical; If FALSE, obj$report() is not run and RTMB variables might not be reported. Default: TRUE.

use.expm

logical; allows to overwrite setting in configuration list (conf$use.expm). If NULL, setting from configuration list is used. Default: NULL.

use.rel.events

logical; allows to overwrite setting in configuration list (conf$use.rel.events). If NULL, setting from configuration list is used. Default: NULL.

dbg

allows to run fit.momo in debugging mode. Default: FALSE.

verbose

if TRUE, print information to console. Default: TRUE.

...

extra arguments to RTMB::MakeADFun.

Value

A list of class momo.fit.

Examples

data(skjepo)

fit <- fit.momo(skjepo, do.sdreport = FALSE)
#> Building the model, that can take a few minutes.
#> Model built (0.8min). Minimizing neg. loglik.
#>   0:     682413.23:  0.00000  0.00000 -4.60517 -4.60517
#>   1:     274130.57: 0.00200891 -0.00117415 -3.69945 -4.18130
#>   2:     132959.07: 0.00197270 -0.00115169 -2.76794 -4.54501
#>   3:     52328.431: 0.00228676 -0.00133323 -1.81275 -4.84101
#>   4:     21296.061: 0.00277104 -0.00161323 -0.981465 -5.39686
#>   5:     9941.5062: 0.00469598 -0.00273318 -0.316932 -6.14411
#>   6:     5014.8354: 0.0326854 -0.0190535 0.287029 -6.94047
#>   7:     3021.4091: 0.442958 -0.257921 0.818323 -7.64214
#>   8:     2373.9455:  1.21661 -0.702660  1.22647 -7.83473
#>   9:     2230.8701:  2.05764 -1.18264  1.47354 -7.87000
#>  10:     2213.2060:  2.92339 -1.67450  1.56594 -7.87448
#>  11:     2203.4125:  5.78530 -3.28752  1.60218 -7.87495
#>  12:     2181.3360:  14.8179 -8.28241  1.64366 -7.87255
#>  13:     2137.0165:  41.1654 -22.4755  1.67783 -7.86106
#>  14:     2107.1179:  67.8229 -36.0774  1.64417 -7.84623
#>  15:     2096.0992:  81.0496 -41.5434  1.58008 -7.83725
#>  16:     2091.6096:  80.5125 -39.2545  1.53774 -7.83667
#>  17:     2089.6137:  79.7125 -37.0433  1.52168 -7.83745
#>  18:     2084.2103:  80.4429 -29.8593  1.48319 -7.84024
#>  19:     2078.7623:  83.8732 -23.5050  1.46125 -7.84291
#>  20:     2058.2771:  101.762 -0.826740  1.40566 -7.85279
#>  21:     2009.6027:  162.051  61.3394  1.30008 -7.88049
#>  22:     1961.4654:  226.671  118.990  1.25059 -7.90713
#>  23:     1870.2841:  371.426  235.847  1.24331 -7.96340
#>  24:     1829.2522:  451.112  297.080  1.33039 -7.99676
#>  25:     1821.5478:  481.850  315.389  1.38495 -8.00712
#>  26:     1820.5711:  488.731  314.998  1.39767 -8.00641
#>  27:     1820.2548:  488.808  309.813  1.40040 -8.00298
#>  28:     1820.2085:  486.448  306.028  1.39900 -8.00064
#>  29:     1820.2053:  485.294  305.007  1.39822 -8.00003
#>  30:     1820.2051:  485.030  304.873  1.39809 -7.99991
#>  31:     1820.2051:  485.018  304.877  1.39809 -7.99985
#>  32:     1820.2051:  485.018  304.877  1.39810 -7.99980
#> Minimization done (0.096min). Model converged. Estimating uncertainty.