Light habitat function

Creates a light habitat class that is being used within momo.

Usage

habi.light(FIELDS, XR, YR, ienv, time.cont)

Arguments

FIELDS

A list with 3-D array fields that are used for interpolation, where the first 2 dimensions span the x and y direction of the spatial field and the third dimension indicates the time dimension.

XR

Limits of the 2-D fields in x direction.

YR

Limits of the 2-D fields in y direction.

ienv

Indicator matrix mapping each model time step to the time steps of each field. The first dimension corresponds to the number of fields and the second to the number of model time steps.

time.cont

Vector with continuous model time steps.

Value

List with functions.