Compare plot

Plotting function that allows to compare multiple fits or fitted and simulated data.

Usage

plotmomo.compare(
  fit,
  ...,
  quantity = c("pref", "taxis", "dif", "par"),
  plot.land = FALSE,
  keep.gpar = FALSE,
  col = momo.cols(10),
  cor.dif = NULL,
  asp = 2,
  plot.legend = 1,
  bg = NULL
)

Arguments

fit

a list of class momo.fit as returned by fit.momo or a list of such (named) fitted objects.

...

optional, additional momo fits.

quantity

quantities that can be compared. The following options are currently implemented: "taxis" for the taxis, "dif" for the diffusion, "par" for the parameters.

plot.land

logical; If TRUE, plot land masses using the function maps::map. Default: FALSE.

keep.gpar

logical; If TRUE, do not overwrite the graphical parameters. Default: FALSE.

col

color

asp

positive number deining the target aspect ratio (columns / rows) of the plot arrangement.

plot.legend

logical or integer indicating which legend should be plotted. Default: 1.

bg

background color of the graphical window. By default (NULL), background is transparent.

cor

scaling parameter for the arrows. By default (NULL), not used.

Value

Nothing.