Plot mark-resight tags

Plot mark-resight tags as lines with sighting positions.

Usage

plotmomo.stags(
  x,
  main = "Mark-resight tags",
  plot.land = FALSE,
  keep.gpar = FALSE,
  xlim = NULL,
  ylim = NULL,
  add = FALSE,
  xlab = "x",
  ylab = "y",
  leg.pos = "topright",
  ...
)

Arguments

x

a list with mark-resight tags or a list of class momo.data, momo.sim, or momo.fit.

main

a main title for the plot. Default: "Mark-resight tags".

plot.land

logical; If TRUE, plot land masses using the function maps::map. Default: FALSE.

keep.gpar

logical; If TRUE, do not overwrite the graphical parameters. Default: FALSE.

xlim

the x limits (x1, x2) of the plot. Note that "x1 > x2" is allowed and leads to a "reversed axis". The default value, NULL, indicates that the range of the finite values to be plotted should be used.

ylim

the y limits of the plot. Default: NULL.

add

logical; if TRUE, no new graphical window is created and tags are added to existing graphical window. Default: FALSE.

xlab

a label for the x axis. Default: "x".

ylab

a label for the y axis. Default: "y".

leg.pos

position of the legend. Default: "topright".

...

additional arguments for the function plot.

Value

Nothing.