Plot taxis

Plot directed advection component (taxis) as arrows on a spatial domain.

Usage

plotmomo.taxis(
  x,
  select = NULL,
  average = TRUE,
  cor = 0.05,
  col = "black",
  alpha = 0.5,
  lwd = 1,
  main = "Taxis",
  plot.land = FALSE,
  keep.gpar = FALSE,
  add = FALSE,
  xlab = "x",
  ylab = "y",
  xaxt = "s",
  yaxt = "s",
  bg = NULL,
  ...
)

Arguments

x

a list of class momo.sim or momo.fit.

select

optional; allows to select specific time steps. Default: NULL.

average

logical; if TRUE (default), plots the average taxis over the selected time steps.

cor

scaling parameter for the arrows. Default: 0.05.

col

color of lines. Default: "black".

alpha

Transparency value. Default: 0.5.

lwd

line width. Default: 1.

main

a main title for the plot. Default: "Taxis".

plot.land

logical; If TRUE, plot land masses using the function maps::map. Default: FALSE.

keep.gpar

logical; If TRUE, do not overwrite the graphical parameters. Default: FALSE.

add

logical; if TRUE, no new graphical window is created and tags are added to existing graphical window. Default: FALSE.

xlab

a label for the x axis. Default: "x".

ylab

a label for the y axis. Default: "y".

xaxt

A character which specifies the x axis type. Specifying ‘"n"’ suppresses plotting of the axis. The standard value is "s": for compatibility with S values "l" and "t" are accepted but are equivalent to "s": any value other than "n" implies plotting.

yaxt

A character which specifies the y axis type. Specifying "n" suppresses plotting.

bg

background color of the graphical window. By default (NULL), background is transparent.

...

additional arguments for the function plot.

Value

Nothing.