Set-up input data for the movement model

setup.momo.data combines all individual data sets, such as tags and environmental data sets and sets up the required data object required to fit the movement model (fit.momo).

Usage

setup.momo.data(
  grid,
  env,
  ctags = NULL,
  atags = NULL,
  stags = NULL,
  effort = NULL,
  knots.tax = NULL,
  knots.dif = NULL,
  const.dif = TRUE,
  trange = NULL,
  dt = NULL,
  time.cont = NULL,
  verbose = TRUE
)

Arguments

grid

a grid object of class momo.grid as returned by the function create.grid.

env

a list with environmental covariates of class momo.env as returned by the function prep.env.

ctags

a data frame with mark-recapture tags of class momo.ctags as returned by the function prep.ctags. Default: NULL.

atags

a list with archival tags of class momo.atags. as returned by the function prep.atags. Default: NULL.

stags

a list with mark-resight tags of class momo.stags. as returned by the function prep.stags. Default: NULL.

effort

a list with effort information of class momo.effort. as returned by the function prep.effort. Default: NULL.

knots.tax

knots for the taxis component. Default: NULL.

knots.dif

knots for the diffusion component. Default: NULL.

const.dif

logical; If TRUE (default), diffusion is assumed to be constant in time and space.

trange

vector of size two defining the limits of the model time range. If NULL (default), time range is defined based on the tags.

dt

time step of the model time. If NULL (default), the time step is set to 0.1.

time.cont

optional; allows to provide a vector representing the model time. Default: NULL.

verbose

if TRUE, print information to console. Default: TRUE.

Value

A list with all data required to fit momo.

Examples

data(skjepo)

ctags <- prep.ctags(skjepo.ctags,
                    names = c("date_time","date_caught",
                              "rel_lon","recap_lon",
                              "rel_lat","recap_lat"),
                    origin = "1899-12-30",
                    speed.limit = 200)
#> Removing 5 recaptured tags for which release time is after recapture time (t0 > t1).

atags <- prep.atags(skjepo.atags,
                    names = c("time","mptlon","mptlat"),
                    origin = "1899-12-30")

grid <- create.grid(grid = skjepo.grid, dxdy = c(10,10))

env <- prep.env(skjepo.env)

dat <- setup.momo.data(grid = grid,
                       env = env,
                       ctags = ctags,
                       atags = atags)