Simulate a simple data set

sim.momo allows to quickly simulate all data sets and components required to fit the movement model, fit.momo.

Usage

sim.momo(
  fit = NULL,
  xrange = c(0, 1),
  yrange = c(0, 1),
  dxdy = c(0.1, 0.1),
  select = FALSE,
  plot.land = FALSE,
  keep.gpar = FALSE,
  trange = c(0, 1),
  dt = 0.1,
  nt = 1,
  rho_t = 0.85,
  sd = 2,
  h = 0.2,
  nu = 2,
  rho_s = 0.8,
  delta = 0.1,
  zrange = c(20, 28),
  matern = TRUE,
  diagonal = FALSE,
  par = NULL,
  n.alpha = 3,
  const.dif = TRUE,
  knots.tax = NULL,
  knots.dif = NULL,
  correct.peclet = FALSE,
  use.ctags = TRUE,
  n.ctags = 500,
  use.atags = TRUE,
  n.atags = 50,
  trange.rel = NULL,
  xrange.rel = c(0.4, 0.6),
  yrange.rel = c(0.4, 0.6),
  verbose = TRUE
)

Arguments

fit

optional; allows to provide a list of class momo.fit to simulate based on the fitted object. By default (NULL), no fitted object is used.

xrange

range of the x-dimension of spatial domain. Default: c(0,1).

yrange

range of the y-dimension of spatial domain. Default: c(0,1).

dxdy

resolution of grid in x and y direction. Default: c(0.1,0.1).

select

logical; if TRUE, allows to select cells in spatial grid. Default: FALSE.

plot.land

logical; If TRUE, plot land masses using the function maps::map. Default: FALSE.

keep.gpar

logical; If TRUE, do not overwrite the graphical parameters. Default: FALSE.

trange

range of model time. Default: c(0,1).

dt

resolution of model time steps. Default: 0.1.

nt

number of environmental fields. Default: 1.

rho_t

autocorrelation coefficient of temporal autocorrelation for environmental fields. Default: 0.85.

sd

standard deviation for simulation of environmental fields. Default: 2.

h

parameter of the matern covariance structure. Default: 0.2.

nu

parameter of the matern covariance structure. Default: 2.

rho_s

parameter of the matern covariance structure. Default: 0.8.

delta

parameter of the matern covariance structure. Default: 0.1.

zrange

range of the environmental covariate. Default: c(20, 28).

matern

logical; if TRUE (default), matern covariance structure is used for simulation of environmental fields.

diagonal

logical; if TRUE, diagonal neighbours are considered in neighbouring structure. Default: FALSE.

par

optional; allows to specify all or some of the parameters used for simulation. Default: NULL.

n.alpha

number of knots for taxis. Default: 3.

const.dif

logical; if TRUE (default), constant diffusion is assumed.

knots.tax

optional; allows to specify knots for the taxis component. Default: NULL.

knots.dif

optional; allows to specify knots for the diffusion component. Default: NULL.

correct.peclet

logical; if TRUE, the diffusion is increased until the peclet number is fulfilled. Default: FALSE.

use.ctags

logical; if TRUE (default), mark-recapture tags are simulated.

n.ctags

number of mark-recapture tags. Default: 500.

use.atags

logical; if TRUE (default), archival tags are simulated.

n.atags

number of archival tags. Default: 50.

trange.rel

optional; allows to specify a specific part of the time series in which tags are released. Default: NULL.

xrange.rel

optional; allows to specify a specific part of the x range of the spatial domain in which tags are released. Default: c(0.4,0.6).

yrange.rel

optional; allows to specify a specific part of the y range of the spatial domain in which tags are released. Default: c(0.4,0.6).

verbose

If TRUE, print information to console. Default: TRUE.

Value

A list of class momo.sim.

Examples

sim <- sim.momo()